+++ controllers/listingskills.js
@@ -12,13 +12,13 @@ const ErrorResponse = require('../utils/errorResponse');
 exports.getListingSkills = asyncHandler(async (req, res, next) => {
   if (req.params.listing_id) {
     // return 404 error response if listing not found
-    const skills = await db.many(
-      'SELECT ls.listing_id, ls.listing_skill_id, ls.skill_id, s.skill FROM ListingSkills ls LEFT JOIN Skills s ON ls.skill_id = s.skill_id WHERE ls.listing_id = $1',
+    const listingSkills = await db.many(
+      'SELECT ls.listing_skill_id, l.listing_id, ls.skill_id, s.skill FROM Listings l LEFT JOIN ListingSkills ls ON l.listing_id = ls.listing_id LEFT JOIN Skills s ON ls.skill_id = s.skill_id WHERE l.listing_id = $1',
       req.params.listing_id
     );
 
     // remove null skill_id from result
-    const data = skills.filter(m => m.milestone_id !== null);
+    const data = listingSkills.filter(ls => ls.listing_skill_id !== null);
 
     return res.status(200).json({
       success: true,
@@ -37,7 +37,7 @@ exports.getListingSkills = asyncHandler(async (req, res, next) => {
  */
 exports.getListingSkill = asyncHandler(async (req, res) => {
   const rows = await db.one(
-    'SELECT * FROM ListingSkills JOIN Skills USING (skill_id) WHERE listing_skill_id = $1',
+    'SELECT * FROM ListingSkills ls JOIN Skills s ON ls.skill_id = s.skill_id WHERE listing_skill_id = $1',
     req.params.id
   );
   res.status(200).json({